# implement transients to cache the API results