--- a/build/aspen-book-list/index-rtl.css
+++ b/build/aspen-book-list/index-rtl.css
@@ -1,11 +1,1 @@
-/*!********************************************************************************************************************************************************************************************************************************************************!*\
-  !*** css ./node_modules/css-loader/dist/cjs.js??ruleSet[1].rules[4].use[1]!./node_modules/postcss-loader/dist/cjs.js??ruleSet[1].rules[4].use[2]!./node_modules/sass-loader/dist/cjs.js??ruleSet[1].rules[4].use[3]!./src/aspen-book-list/editor.scss ***!
-  \********************************************************************************************************************************************************************************************************************************************************/
-/**
- * The following styles get applied inside the editor only.
- *
- * Replace them with your own styles or remove the file completely.
- */
-.wp-block-create-block-aspen-book-list {
-  border: 1px dotted #f00;
-}
+.wp-block-create-block-aspen-book-list{border:1px dotted red}