--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "aspen-book-list",
-	"version": "0.2.0",
+	"version": "0.2.1",
 	"description": "Display a list from the Aspen Discovery API",
 	"author": "Cleveland Public Library (Will skora and Tyrone Fontaine)",
 	"license": "GPL-2.0-or-later",