--- a/src/aspen-book-list/block.json
+++ b/src/aspen-book-list/block.json
@@ -2,7 +2,7 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 3,
 	"name": "create-block/aspen-book-list",
-	"version": "0.2.0",
+	"version": "0.2.1",
 	"title": "Aspen Book List",
 	"category": "widgets",
 	"icon": "book",