This caches the api results after a list has been fetched from Aspen's api for the first time. 

The list will remain in WordPress' cache for 7 days. Note that if there are any updates to a list in Aspen and a list already has been loaded onto the website's content, then the updated list will not appear for up to 7 days in the block. 

There are ways to manually delete the transient, which will force a new API call to the Aspen server, returning updated results. 
(e.g. `wp transient delete aspen_api_list_id_###` in the server (replace ### with the list id from Aspen). 

remarkably improved site performance: 
before: 
Page Generation Time
2.3346s
3.9% of 60s limit

Peak Memory Usage
21,238,232 bytes (20.3 MB)
2.5% of 800 MB server limit

Database Queries
0.0105s

Total: 93

HTTP API Calls
1.8194s

Total: 4

AFTER: 

Page Generation Time
0.3196s
0.5% of 60s limit

Peak Memory Usage
18,879,848 bytes (18.0 MB)
2.3% of 800 MB server limit

Database Queries
0.0064s

Total: 97

HTTP API Calls
None